refactor: single-source config, fix stale constants (#29)
Two config modules had diverged (g_nfl/__init__.py vs utils/config.py):
HFA 1.5 vs 1.3, plus duplicated AVG_POINTS/thresholds and a dead WP_FILTER.
Consolidate to utils/config.py as the single source; point the two root
importers at it and empty __init__.py.

Fix stale values: SPREAD_STDEV 11.5 -> 13.5 (realized home-margin std is ~14,
~13.3 trimmed; the old value made percentile->spread too narrow). Keep live
HFA=1.5 (config's 1.3 was dead). Update CLAUDE.md to match.
