--- src/g_nfl/__init__.py
@@ -1,22 +1,0 @@
-CUR_WEEK = 12
-CUR_SEASON = 2025
-
-HFA = 1.5
-WP_FILTER = 0.025
-
-AVG_POINTS = 21.5
-SPREAD_STDEV = 11.5
-
-EXPLOSIVE_RUN_THRESHOLD = 10
-EXPLOSIVE_PASS_THRESHOLD = 15
-
-__all__ = [
-    "CUR_WEEK",
-    "CUR_SEASON",
-    "HFA",
-    "WP_FILTER",
-    "EXPLOSIVE_RUN_THRESHOLD",
-    "EXPLOSIVE_PASS_THRESHOLD",
-    "AVG_POINTS",
-    "SPREAD_STDEV",
-]
